# docs/openapi.yaml
openapi: 3.0.3
info:
  title: RideAuto API (публичные эндпоинты)
  version: "1.0"
  description: |
    Минимальное описание read-only API каталога и служебных маршрутов.
    Полный набор query-параметров для `/api/cars` и ответы см. в `backend/fastapi_app/routers/search.py` и схемах Pydantic.
    Лимиты GET по IP: `WRA_RATE_LIMIT_GET_CARS_PER_MINUTE`, `WRA_RATE_LIMIT_GET_FACETS_PER_MINUTE` (0 = без лимита).
    Курсорная пагинация: query `cursor` в `/api/cars` для сортировок date_*, year_*, mileage_*, price_high / price_low; см. `meta.next_cursor`.
    Без `cursor`: глубокий `page` ограничен для анонимов (`WRA_CATALOG_MAX_PAGE_OFFSET_ANON`, по умолчанию 80; `-1` = выкл.); при сессии Bearer — `WRA_CATALOG_MAX_PAGE_OFFSET_AUTH` (0 = без лимита).
    Кэш фасетов в воркере: `WRA_FACETS_CACHE_TTL` (сек, по умолчанию 300).
    Метрики: `GET /metrics` (legacy alias: `GET /api/metrics`) при `WRA_METRICS_ENABLED=1`
    (legacy env alias: `WRA_PROMETHEUS_METRICS=1`).
servers:
  - url: /
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
    cookieAuth:
      type: apiKey
      in: cookie
      name: wra_session
      description: Сессия magic-link (WRA_AUTH_COOKIE_NAME, по умолчанию wra_session)
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: unauthorized
...
paths:
...
  /api/car/{id}:
    get:
      summary: Одна карточка по id
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          content:
            application/json:
              schema: { type: object }
        "404":
          description: Не найдено
  /api/compare:
    get:
      summary: Краткие поля для сравнения (до 4 id через запятую)
      parameters:
        - name: ids
          in: query
          description: "Например: c1,c2,c3"
          schema: { type: string }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: array
                    items: { type: object }
...
  /api/meili/outbox/process:
    post:
      summary: Обработать PG→Meili outbox (X-Admin-Key)
      parameters:
        - name: X-Admin-Key
          in: header
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 200 }
      responses:
        "200":
          content:
            application/json:
              schema: { type: object }
        "401":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /api/logout:
    post:
      summary: Выход (инвалидировать сессию на сервере)
      security:
        - bearerAuth: []
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
  /api/me:
    get:
      summary: Текущий пользователь
      security:
        - bearerAuth: []
      responses:
        "200":
          content:
            application/json:
              schema: { type: object }
        "401":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/favorites:
    get:
      summary: Избранное
      security:
        - bearerAuth: []
      responses:
        "200":
          content:
            application/json:
              schema: { type: object }
        "401":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    post:
      summary: Добавить в избранное
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                car_id: { type: string }
                note: { type: string, nullable: true }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
        "400":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/favorites/{car_id}:
    delete:
      summary: Удалить из избранного
      security:
        - bearerAuth: []
      parameters:
        - name: car_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
        "401":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...